Bug修复 P0 ProcessRunner 死锁、子进程未终止、异常吞噬、GooglePlay 异常覆盖、ConfigEditor 白名单不同步 P1 Android 管线步骤条件、BuildNumber 类型一致性、配置校验前置、符号链接复制、clone 恢复、dry-run 预览、HttpClient 超时、RunId 精度 P2 Console 输出清理、Path.GetDirectoryName 空引用、null 安全、JsonSerializerOptions 复用、Unity 路径占位符、平台切换清理、冗余检查 测试覆盖：193 个测试全部通过

+++ Configuration/BuildConfig.Android.cs
@@ -15,7 +15,7 @@ internal sealed partial class BuildConfig
         }
 
         if (!string.IsNullOrWhiteSpace(BuildNumber) &&
-            (!int.TryParse(BuildNumber, out int versionCode) || versionCode <= 0))
+            (!long.TryParse(BuildNumber, out long versionCode) || versionCode <= 0))
         {
             throw new InvalidOperationException("Android buildNumber/versionCode 必须是大于 0 的整数。");
         }
